requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pandas
openpyxl
numpy
python-magic
requests
pyTelegramBotAPI
matplotlib